code review changes - split program.cs
